Extract the logic into the seperate file
- allows to unit test same function as in `wrap_go.c`
