+++ src/wrap_go.c
@@ -15,6 +15,7 @@
 #include "state.h"
 #include "utils.h"
 #include "fn.h"
+#include "oci.h"
 #include "capstone/capstone.h"
 #include "scopestdlib.h"
 #include "snapshot.h"
@@ -460,363 +461,6 @@ free_go_str(char *str) {
 }
 */
 
-/*
- * Rewrite the container configuration for the given container.
- * A path to the container specific work dir defines the location of the configuration file.
- *
- * Please look into opencontainers Linux runtime-spec for details about the exact JSON struct.
- * The following changes will be performed:
- * - Add a mount points
- *   `appscope` directory will be mounted from the host "/usr/lib/appscope/" into the container: "/usr/lib/appscope/"
- *   
- *   UNIX socket directory will be mounted from the host into the container the path to UNIX socket will be read from
- *   host based on value in the filter file
- *
- * - Extend Environment variables
- *   `LD_PRELOAD` will contain the following entry `/opt/appscope/libscope.so`
- *   `SCOPE_SETUP_DONE=true` mark that configuration was processed
- *
- * - Add prestart hook
- *   execute scope extract operation to ensure using library with proper loader reference (musl/glibc)
- */
-static void
-rewriteOpenContainersConfig(const char *cWorkDir)
-{
-    char path[PATH_MAX] = {0};
-    char scopePath[PATH_MAX] = {0};
-
-    if (!cWorkDir) {
-        goto exit;
-    }
-
-    if (scope_snprintf(path, sizeof(path), "%s/config.json", cWorkDir) < 0) {
-        goto exit;
-    }
-
-    struct stat fileStat;
-    if (scope_stat(path, &fileStat) == -1) {
-        goto exit;
-    }
-
-    FILE *fp = scope_fopen(path, "r");
-    if (!fp) {
-        goto exit;
-    }
-
-    // TODO: enable this
-    // if (scope_stat("/usr/lib/appscope/scope_filter", &fileStat) == -1) {
-    //     scope_fclose(fp);
-    //     goto exit;
-    // }
-
-    // Scope executable must exists
-    if (scope_snprintf(scopePath, sizeof(scopePath), "/usr/lib/appscope/%s/scope", libVersion(SCOPE_VER)) < 0) {
-        scope_fclose(fp);
-        goto exit;
-    }
-
-    if (scope_stat(scopePath, &fileStat) == -1) {
-        goto exit;
-    }
-
-    /*
-    * Read the file contents into a string
-    */
-    char *buf = (char *)scope_malloc(fileStat.st_size);
-    if (!buf) {
-        scope_fclose(fp);
-        goto exit;
-    }
-
-    scope_fread(buf, sizeof(char), fileStat.st_size, fp);
-    scope_fclose(fp);
-
-    cJSON *json = cJSON_Parse(buf);
-    scope_free(buf);
-    if (json == NULL) {
-        goto exit;
-    }
-
-    /*
-    * Handle process environment variables
-    *
-    "env":[
-         "PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin",
-         "HOSTNAME=6735578591bb",
-         "TERM=xterm",
-         "LD_PRELOAD=/opt/appscope/libscope.so",
-         "SCOPE_SETUP_DONE=true"
-      ],
-    */
-    cJSON *procNode = cJSON_GetObjectItemCaseSensitive(json, "process");
-    if (!procNode) {
-        procNode = cJSON_CreateObject();
-        if (!procNode) {
-            cJSON_Delete(json);
-            goto exit;
-        }
-        cJSON_AddItemToObject(json, "process", procNode);
-    }
-
-    cJSON *envNodeArr = cJSON_GetObjectItemCaseSensitive(procNode, "env");
-    if (envNodeArr) {
-        bool ldPreloadPresent = FALSE;
-        // Iterate over environment string array
-        size_t envSize = cJSON_GetArraySize(envNodeArr);
-        for (int i = 0; i < envSize ;++i) {
-            cJSON *item = cJSON_GetArrayItem(envNodeArr, i);
-            char *strItem = cJSON_GetStringValue(item);
-
-            if (scope_strncmp("LD_PRELOAD=", strItem, C_STRLEN("LD_PRELOAD=")) == 0) {
-                size_t itemLen = scope_strlen(strItem);
-                size_t newLdprelLen = itemLen + C_STRLEN("/opt/appscope/libscope.so:");
-                char *newLdPreloadLib = scope_calloc(1, newLdprelLen);
-                if (!newLdPreloadLib) {
-                    cJSON_Delete(json);
-                    goto exit;
-                }
-                scope_strncpy(newLdPreloadLib, "LD_PRELOAD=/opt/appscope/libscope.so:", C_STRLEN("LD_PRELOAD=/opt/appscope/libscope.so:"));
-                scope_strcat(newLdPreloadLib, strItem + C_STRLEN("LD_PRELOAD="));
-                cJSON *newLdPreloadLibObj = cJSON_CreateString(newLdPreloadLib);
-                if (!newLdPreloadLibObj) {
-                    scope_free(newLdPreloadLib);
-                    cJSON_Delete(json);
-                    goto exit;
-                }
-                cJSON_ReplaceItemInArray(envNodeArr, i, newLdPreloadLibObj);
-                scope_free(newLdPreloadLib);
-
-                cJSON *scopeEnvNode = cJSON_CreateString("SCOPE_SETUP_DONE=true");
-                if (!scopeEnvNode) {
-                    cJSON_Delete(json);
-                    goto exit;
-                }
-                cJSON_AddItemToArray(envNodeArr, scopeEnvNode);
-                ldPreloadPresent = TRUE;
-                break;
-            } else if (scope_strncmp("SCOPE_SETUP_DONE=true", strItem, C_STRLEN("SCOPE_SETUP_DONE=true")) == 0) {
-                // we are done here
-                cJSON_Delete(json);
-                goto exit;
-            }
-        }
-
-
-        // There was no LD_PRELOAD in environment variables
-        if (ldPreloadPresent == FALSE) {
-            const char *const envItems[2] =
-            {
-                "LD_PRELOAD=/opt/appscope/libscope.so",
-                "SCOPE_SETUP_DONE=true"
-            };
-            for (int i = 0; i < ARRAY_SIZE(envItems) ;++i) {
-                cJSON *scopeEnvNode = cJSON_CreateString(envItems[i]);
-                if (!scopeEnvNode) {
-                    cJSON_Delete(json);
-                    goto exit;
-                }
-                cJSON_AddItemToArray(envNodeArr, scopeEnvNode);
-            }
-        }
-    } else {
-        const char * envItems[2] =
-        {
-            "LD_PRELOAD=/opt/appscope/libscope.so",
-            "SCOPE_SETUP_DONE=true"
-        };
-        envNodeArr = cJSON_CreateStringArray(envItems, ARRAY_SIZE(envItems));
-        if (!envNodeArr) {
-            cJSON_Delete(json);
-            goto exit;
-        }
-        cJSON_AddItemToObject(procNode, "env", envNodeArr);
-    }
-
-    /*
-    * Handle process mounts for library and filter file and socket
-    *
-    "mounts":[
-      {
-         "destination":"/proc",
-         "type":"proc",
-         "source":"proc",
-         "options":[
-            "nosuid",
-            "noexec",
-            "nodev"
-         ]
-      },
-      ...
-      {
-         "destination":"/usr/lib/appscope/",
-         "type":"bind",
-         "source":"/usr/lib/appscope/",
-         "options":[
-            "rbind",
-            "rprivate"
-         ]
-      },
-      {
-         "destination":"/var/run/appscope/",
-         "type":"bind",
-         "source":"/var/run/appscope/",
-         "options":[
-            "rbind",
-            "rprivate"
-         ]
-      }
-    */
-
-    const char *mountPath[2] =
-    {
-        "/usr/lib/appscope/",
-        "/var/run/appscope/"
-    };
-
-    for (int i = 0; i < ARRAY_SIZE(mountPath); ++i ) {
-        cJSON *mountNodeArr = cJSON_GetObjectItemCaseSensitive(json, "mounts");
-        if (!mountNodeArr) {
-            mountNodeArr = cJSON_CreateArray();
-            if (!mountNodeArr) {
-                cJSON_Delete(json);
-                goto exit;
-            }
-            cJSON_AddItemToObject(json, "mounts", mountNodeArr);
-        }
-
-        cJSON *mountNode = cJSON_CreateObject();
-        if (!mountNode) {
-            cJSON_Delete(json);
-            goto exit;
-        }
-
-        if (!cJSON_AddStringToObjLN(mountNode, "destination", mountPath[i])) {
-            cJSON_Delete(mountNode);
-            cJSON_Delete(json);
-            goto exit;
-        }
-
-        if (!cJSON_AddStringToObjLN(mountNode, "type", "bind")) {
-            cJSON_Delete(mountNode);
-            cJSON_Delete(json);
-            goto exit;
-        }
-
-        if (!cJSON_AddStringToObjLN(mountNode, "source", mountPath[i])) {
-            cJSON_Delete(mountNode);
-            cJSON_Delete(json);
-            goto exit;
-        }
-
-        const char *optItems[2] =
-        {
-            "rbind",
-            "rprivate"
-        };
-
-        cJSON *optNodeArr = cJSON_CreateStringArray(optItems, ARRAY_SIZE(optItems));
-        if (!optNodeArr) {
-            cJSON_Delete(mountNode);
-            cJSON_Delete(json);
-            goto exit;
-        }
-        cJSON_AddItemToObject(mountNode, "options", optNodeArr);
-        cJSON_AddItemToArray(mountNodeArr, mountNode);
-    }
-
-    /*
-    * Handle startContainer hooks process
-    *
-   "hooks":{
-      "prestart":[
-         {
-            "path":"/proc/1513/exe",
-            "args":[
-               "libnetwork-setkey",
-               "-exec-root=/var/run/docker",
-               "6735578591bb3c5aebc91e5c702470c52d2c10cea52e4836604bf5a4a6c0f2eb",
-               "ec7e49ffc98c"
-            ]
-         }
-      ],
-      "startContainer":[
-         {
-            "path":"/usr/lib/appscope/<version>/scope"
-            "args":[
-               "/usr/lib/appscope/<version>/scope",
-               "extract",
-               "-p",
-               "/opt/appscope",
-            ]
-         },
-       ]
-    */
-    cJSON *hooksNode = cJSON_GetObjectItemCaseSensitive(json, "hooks");
-    if (!hooksNode) {
-        hooksNode = cJSON_CreateObject();
-        if (!hooksNode) {
-            cJSON_Delete(json);
-            goto exit;
-        }
-        cJSON_AddItemToObject(json, "hooks", hooksNode);
-    }
-
-    cJSON *startContainerNodeArr = cJSON_GetObjectItemCaseSensitive(hooksNode, "startContainer");
-    if (!startContainerNodeArr) {
-        startContainerNodeArr = cJSON_CreateArray();
-        if (!startContainerNodeArr) {
-            cJSON_Delete(json);
-            goto exit;
-        }
-        cJSON_AddItemToObject(hooksNode, "startContainer", startContainerNodeArr);
-    }
-
-    cJSON *startContainerNode = cJSON_CreateObject();
-    if (!startContainerNode) {
-        cJSON_Delete(json);
-        goto exit;
-    }
-
-    if (!cJSON_AddStringToObjLN(startContainerNode, "path",  scopePath)) {
-        cJSON_Delete(startContainerNode);
-        cJSON_Delete(json);
-        goto exit;
-    }
-
-    const char *argsItems[4] =
-    {
-        scopePath,
-        "extract",
-        "-p",
-        "/opt/appscope"
-    };
-    cJSON *argsNodeArr = cJSON_CreateStringArray(argsItems, ARRAY_SIZE(argsItems));
-    if (!argsNodeArr) {
-        cJSON_Delete(startContainerNode);
-        cJSON_Delete(json);
-        goto exit;
-    }
-    cJSON_AddItemToObject(startContainerNode, "args", argsNodeArr);
-    cJSON_AddItemToArray(startContainerNodeArr, startContainerNode);
-
-    char *jsonStr = cJSON_PrintUnformatted(json);
-    cJSON_Delete(json);
-
-    // Overwrite the file
-    fp = scope_fopen(path, "w");
-    if (fp == NULL) {
-        cJSON_free(jsonStr);
-        goto exit;
-    }
-
-    scope_fprintf(fp, "%s\n", jsonStr);
-
-    cJSON_free(jsonStr);
-    scope_fclose(fp);
-
-exit:
-    return;
-}
 
 static void
 containerStart(void)
@@ -849,7 +493,41 @@ containerStart(void)
         }
     }
 
-    if (cWorkDir) rewriteOpenContainersConfig(cWorkDir);
+    if (cWorkDir) {
+        char cfgPath[PATH_MAX] = {0};
+        char scopePath[PATH_MAX] = {0};
+        struct stat fileStat;
+
+        if (scope_snprintf(scopePath, sizeof(scopePath), "/usr/lib/appscope/%s/scope", libVersion(SCOPE_VER)) < 0) {
+            goto exit;
+        }
+
+        if (scope_stat(scopePath, &fileStat) == -1) {
+            sysprint("\t%s: scope is not accessible %s\n", __FUNCTION__, scopePath);
+            goto exit;
+        }
+
+        if (scope_snprintf(cfgPath, sizeof(cfgPath), "%s/config.json", cWorkDir) < 0) {
+            goto exit;
+        }
+
+        void *cfgMem = ociReadCfgIntoMem(cfgPath);
+        if (!cfgMem) {
+            goto exit;
+        }
+
+        char *modCfgMem = ociModifyCfg(cfgMem, scopePath);
+        scope_free(cfgMem);
+
+        if (modCfgMem) {
+            ociWriteConfig(cfgPath, modCfgMem);
+            scope_free(modCfgMem);
+        } else {
+            sysprint("\t%s: Modify OCI config fails config: %s, scope: %s,  \n", __FUNCTION__, cfgPath, scopePath);
+        }
+    }
+
+exit:
     scope_free(buf);
 }
