+++ test/unit/library/ocitest.c
@@ -2,16 +2,10 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include "oci.h"
 #include "cJSON.h"
 #include "test.h"
 #include "scopestdlib.h"
-
-#ifndef FALSE
-#define FALSE   0
-#endif
-#ifndef TRUE
-#define TRUE   1
-#endif
 
 static char dirPath[PATH_MAX];
 
@@ -51,275 +45,68 @@ testDirPath(char *path, const char *argv0) {
     return 0;
 }
 
-// This function reflects the logic provided by rewriteOpenContainersConfig
 static bool
-rewriteOpenContainersConfigTest(int id) {
-    char inPath [PATH_MAX] = {0};
+verifyModifiedCfg(int id, const char* cmpStr) {
     char outPath [PATH_MAX] = {0};
-    scope_snprintf(inPath, PATH_MAX, "%s/data/oci/oci%din.json", dirPath, id);
     scope_snprintf(outPath, PATH_MAX, "%s/data/oci/oci%dout.json", dirPath, id);
 
-    const char * scopeWithVersion = "/usr/lib/appscope/1.2.3/scope";
-    bool res = FALSE;
-    struct stat fileStat;
-    if (scope_stat(inPath, &fileStat) == -1) {
-        assert_non_null(NULL);
-        return res;
-    }
-
-    FILE *fp = scope_fopen(inPath, "r");
-    if (!fp) {
-        assert_non_null(NULL);
-        return res;
-    }
-
-    /*
-    * Read the file contents into a string
-    */
-    char *buf = (char *)scope_malloc(fileStat.st_size);
-    if (!buf) {
-        scope_fclose(fp);
-        assert_non_null(NULL);
-        return res;
-    }
-
-    scope_fread(buf, sizeof(char), fileStat.st_size, fp);
-    scope_fclose(fp);
-
-
-    cJSON *json = cJSON_Parse(buf);
-    if (!json) {
-        assert_non_null(NULL);
-        return res;
-    }
-    scope_free(buf);
-    char *jsonStr = NULL;
-
-    // Handle the process
-    cJSON *procNode = cJSON_GetObjectItemCaseSensitive(json, "process");
-    if (!procNode) {
-        procNode = cJSON_CreateObject();
-        if (!procNode) {
-            assert_non_null(NULL);
-            goto exit;
-        }
-        cJSON_AddItemToObject(json, "process", procNode);
-    }
-    cJSON *envNodeArr = cJSON_GetObjectItemCaseSensitive(procNode, "env");
-    if (envNodeArr) {
-        bool ldPreloadPresent = FALSE;
-        // Iterate over environment string array
-        size_t envSize = cJSON_GetArraySize(envNodeArr);
-        for (int i = 0; i < envSize ;++i) {
-            cJSON *item = cJSON_GetArrayItem(envNodeArr, i);
-            char *strItem = cJSON_GetStringValue(item);
-
-            if (scope_strncmp("LD_PRELOAD=", strItem, C_STRLEN("LD_PRELOAD=")) == 0) {
-                size_t itemLen = scope_strlen(strItem);
-                size_t newLdprelLen = itemLen + C_STRLEN("/opt/appscope/libscope.so:");
-                char *newLdPreloadLib = scope_calloc(1, newLdprelLen);
-                if (!newLdPreloadLib) {
-                    assert_non_null(NULL);
-                    goto exit;
-                }
-                scope_strncpy(newLdPreloadLib, "LD_PRELOAD=/opt/appscope/libscope.so:", C_STRLEN("LD_PRELOAD=/opt/appscope/libscope.so:"));
-                scope_strcat(newLdPreloadLib, strItem + C_STRLEN("LD_PRELOAD="));
-                cJSON *newLdPreloadLibObj = cJSON_CreateString(newLdPreloadLib);
-                if (!newLdPreloadLibObj) {
-                    scope_free(newLdPreloadLib);
-                    assert_non_null(NULL);
-                    goto exit;
-                }
-                cJSON_ReplaceItemInArray(envNodeArr, i, newLdPreloadLibObj);
-                scope_free(newLdPreloadLib);
-
-                cJSON *scopeEnvNode = cJSON_CreateString("SCOPE_SETUP_DONE=true");
-                if (!scopeEnvNode) {
-                    assert_non_null(NULL);
-                    goto exit;
-                }
-                cJSON_AddItemToArray(envNodeArr, scopeEnvNode);
-                ldPreloadPresent = TRUE;
-                break;
-            } else if (scope_strncmp("SCOPE_SETUP_DONE=true", strItem, C_STRLEN("SCOPE_SETUP_DONE=true")) == 0) {
-                // we are done here
-                res = TRUE;
-                goto exit;
-            }
-        }
-
-
-        // There was no LD_PRELOAD in environment variables
-        if (ldPreloadPresent == FALSE) {
-            const char *const envItems[2] =
-            {
-                "LD_PRELOAD=/opt/appscope/libscope.so",
-                "SCOPE_SETUP_DONE=true"
-            };
-            for (int i = 0; i < ARRAY_SIZE(envItems) ;++i) {
-                cJSON *scopeEnvNode = cJSON_CreateString(envItems[i]);
-                if (!scopeEnvNode) {
-                    assert_non_null(NULL);
-                    goto exit;
-                }
-                cJSON_AddItemToArray(envNodeArr, scopeEnvNode);
-            }
-        }
-    } else {
-        const char *envItems[2] =
-        {
-            "LD_PRELOAD=/opt/appscope/libscope.so",
-            "SCOPE_SETUP_DONE=true"
-        };
-        envNodeArr = cJSON_CreateStringArray(envItems, ARRAY_SIZE(envItems));
-        if (!envNodeArr) {
-            assert_non_null(NULL);
-            goto exit;
-        }
-        cJSON_AddItemToObject(procNode, "env", envNodeArr);
-    }
-
-    const char *mountPath[2] =
-    {
-        "/usr/lib/appscope/",
-        "/var/run/appscope/"
-    };
-
-    for (int i = 0; i < 2; ++i ) {
-        cJSON *mountNodeArr = cJSON_GetObjectItemCaseSensitive(json, "mounts");
-        if (!mountNodeArr) {
-            mountNodeArr = cJSON_CreateArray();
-            if (!mountNodeArr) {
-                assert_non_null(NULL);
-                goto exit;
-            }
-            cJSON_AddItemToObject(json, "mounts", mountNodeArr);
-        }
-
-        cJSON *mountNode = cJSON_CreateObject();
-        if (!mountNode) {
-            assert_non_null(NULL);
-            goto exit;
-        }
-
-        if (!cJSON_AddStringToObjLN(mountNode, "destination", mountPath[i])) {
-            cJSON_Delete(mountNode);
-            assert_non_null(NULL);
-            goto exit;
-        }
-
-        if (!cJSON_AddStringToObjLN(mountNode, "type", "bind")) {
-            cJSON_Delete(mountNode);
-            assert_non_null(NULL);
-            goto exit;
-        }
-
-        if (!cJSON_AddStringToObjLN(mountNode, "source", mountPath[i])) {
-            cJSON_Delete(mountNode);
-            assert_non_null(NULL);
-            goto exit;
-        }
-
-        const char *optItems[2] =
-        {
-            "rbind",
-            "rprivate"
-        };
-
-        cJSON *optNodeArr = cJSON_CreateStringArray(optItems, ARRAY_SIZE(optItems));
-        if (!optNodeArr) {
-            cJSON_Delete(mountNode);
-            assert_non_null(NULL);
-            goto exit;
-        }
-        cJSON_AddItemToObject(mountNode, "options", optNodeArr);
-        cJSON_AddItemToArray(mountNodeArr, mountNode);
-    }
-
-    cJSON *hooksNode = cJSON_GetObjectItemCaseSensitive(json, "hooks");
-    if (!hooksNode) {
-        hooksNode = cJSON_CreateObject();
-        if (!hooksNode) {
-            assert_non_null(NULL);
-            goto exit;
-        }
-        cJSON_AddItemToObject(json, "hooks", hooksNode);
-    }
-
-    cJSON *startContainerNodeArr = cJSON_GetObjectItemCaseSensitive(hooksNode, "startContainer");
-    if (!startContainerNodeArr) {
-        startContainerNodeArr = cJSON_CreateArray();
-        if (!startContainerNodeArr) {
-            assert_non_null(NULL);
-            goto exit;
-        }
-        cJSON_AddItemToObject(hooksNode, "startContainer", startContainerNodeArr);
-    }
-
-    cJSON *startContainerNode = cJSON_CreateObject();
-    if (!startContainerNode) {
-        assert_non_null(NULL);
-        goto exit;
-    }
-
-    if (!cJSON_AddStringToObjLN(startContainerNode, "path",  scopeWithVersion)) {
-        cJSON_Delete(startContainerNode);
-        assert_non_null(NULL);
-        goto exit;
-    }
-
-    const char *argsItems[4] =
-    {
-        scopeWithVersion,
-        "extract",
-        "-p",
-        "/opt/appscope"
-    };
-    cJSON *argsNodeArr = cJSON_CreateStringArray(argsItems, ARRAY_SIZE(argsItems));
-    if (!argsNodeArr) {
-        cJSON_Delete(startContainerNode);
-        assert_non_null(NULL);
-        goto exit;
-    }
-    cJSON_AddItemToObject(startContainerNode, "args", argsNodeArr);
-    cJSON_AddItemToArray(startContainerNodeArr, startContainerNode);
-
-    jsonStr = cJSON_Print(json);
-
-    // Overwrite the file
     int fdOut = scope_open(outPath, O_RDONLY);
     if (fdOut == -1) {
-        cJSON_free(jsonStr);
-        goto exit;
+        assert_non_null(NULL);
+        return FALSE;
     }
 
     struct stat stOut;
     if (scope_fstat(fdOut, &stOut) == -1) {
-        cJSON_free(jsonStr);
+        assert_non_null(NULL);
         scope_close(fdOut);
-        goto exit;
+        return FALSE;
     }
-
 
     void *fdOutMap = scope_mmap(NULL, stOut.st_size, PROT_READ, MAP_PRIVATE, fdOut, 0);
     if (fdOutMap == MAP_FAILED) {
-        cJSON_free(jsonStr);
-        scope_close(fdOut);
-        goto exit;
-    }
-
-    if (memcmp(fdOutMap, jsonStr, stOut.st_size) != 0) {
-        scope_fprintf(scope_stderr, jsonStr);
         assert_non_null(NULL);
+        scope_close(fdOut);
+        return FALSE;
     }
-    scope_munmap(fdOutMap, stOut.st_size);
-
-    cJSON_free(jsonStr);
     scope_close(fdOut);
 
-    res = TRUE;
-exit:
-    cJSON_Delete(json);
+    cJSON* parseOut = cJSON_Parse(fdOutMap);
+    char* outBuf = cJSON_PrintUnformatted(parseOut);
+    cJSON_Delete(parseOut);
+
+    assert_int_equal(scope_strlen(outBuf), scope_strlen(cmpStr));
+    if (scope_strcmp(cmpStr, outBuf) != 0 ){
+        scope_fprintf(scope_stderr, cmpStr);
+        scope_fprintf(scope_stderr, outBuf);
+        assert_non_null(NULL);
+        scope_munmap(fdOutMap, stOut.st_size);
+        return FALSE;
+    }
+
+    scope_free(outBuf);
+    scope_munmap(fdOutMap, stOut.st_size);
+
+    return TRUE;
+}
+
+// This function reflects the logic provided by rewriteOpenContainersConfig
+static bool
+rewriteOpenContainersConfigTest(int id) {
+
+    char inPath [PATH_MAX] = {0};
+    scope_snprintf(inPath, PATH_MAX, "%s/data/oci/oci%din.json", dirPath, id);
+    const char * scopeWithVersion = "/usr/lib/appscope/1.2.3/scope";
+
+    void *cfgMem = ociReadCfgIntoMem(inPath);
+
+    char* modifMem = ociModifyCfg(cfgMem, scopeWithVersion);
+
+    bool res = verifyModifiedCfg(id, modifMem);
+
+    scope_free(cfgMem);
+    scope_free(modifMem);
+
     return res;
 }
 
@@ -375,7 +162,7 @@ ocitest_hooks_complete(void **state)
 int
 main(int argc, char* argv[])
 {
-    printf("running %s\n", argv[0]);
+    scope_printf("running %s\n", argv[0]);
     if (testDirPath(dirPath, argv[0])) {
         return EXIT_FAILURE;
     }
